requirements.txt
pdfplumber>=0.10.0
google-genai>=1.0.0
pydantic>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pytest>=8.0.0
requests>=2.31.0
